+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- location query based discovery
 - requirement test execution logging
 - test for glue:R0056
 - configuration option to change the maximum waiting time during connection establishment